Update CLion Program arguments on tree update
I ended up hooking into the current/virtual workspace.xml and I then check if we have the program arguments XML tag for `CMakeRunConfiguration` and if not we add one with the current CLArgs, if we do find the tag `PROGRAM_PARAMS` we instead do some basic pattern matching and replace the old CLArgs with the new ones.

For now we update the Program arguments on tree change, it may make sense to do this on program run, there is something similar for cmake build for running code before cmake build so there may be something similar for run

I wanted to avoid parsing the XML code initially but I couldn't find the code for updating the current projects program arguments, I'm sure the open API call exists but the jetbrains documents is basically non existent for CLion.

It seems that for Rider C++ projects that use .vcxproj there is just a simple command line patcher you can call but I couldn't find a similar call for CLion

All of this code took way longer to work out that I hoped as I wasn't aware of Java Virtual files before this. thankfully jetbrains provide a virtual file util functions. from my understanding these calls flush any changes to a selected VF and makes the IDE aware of file changes which makes all of this possible, it took quite some time to find the function calls I needed to make the final stage of this plugin even possible.

the resulting code ended up being pretty small in the end, but it is really in need of some bug crunching there are a few UI issues which get on my nerves while using the plugin which I would like to address before making this all public too.